Fix CreateMenu and mobile safari attachment picker
